ooooooooooooo

+++ transmitter/types.go
@@ -6,7 +6,7 @@ package transmitter
 // 2: received, from Rx to Tx
 // 3: falied to receive, from Rx to Tx
 type Signal struct {
-	Index  int64  `json:"i"`
+	Index  int32  `json:"i"`
 	Action uint8  `json:"a"`
 	ErrS   string `json:"e"`
 }
